backend/requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.0,<5.0
gunicorn>=22.0,<23.0
Pillow>=11.0,<12.0
psycopg[binary]>=3.1,<4.0
